requirements.txt
gunicorn==20.1.0
click==8.0.4
distro==1.8.0
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
opencv-contrib-python==4.7.0.68
opencv-python==4.7.0.68
packaging==21.3
Pillow==8.4.0
pycodestyle==2.10.0
pyparsing==3.0.9
scikit-build==0.16.4
tomli==1.2.3
typing_extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
